+++ premium/packages/preact-scheduler/src/resource-timeline/components/spreadsheet/ResourceSubrow.tsx
@@ -78,6 +78,7 @@ export class ResourceSubrow extends BaseComponent<ResourceSubrowProps, ViewConte
               key={i} // eslint-disable-line react/no-array-index-key
               colSpec={colSpec}
               resource={resource}
+              field={colSpec.field || 'title'}
               fieldValue={fieldValue}
               indent={props.indent}
               hasChildren={props.hasChildren}